Moved Comet related utility methods into `trainer/utils.py` to resolve merge conflict with master branch,
