+++ trl/trainer/bco_trainer.py
@@ -53,13 +53,13 @@ from transformers.utils import is_peft_available
 
 from ..data_utils import maybe_apply_chat_template
 from ..models import PreTrainedModelWrapper, create_reference_model
-from ..utils import get_comet_experiment_url
 from .bco_config import BCOConfig
 from .utils import (
     DPODataCollatorWithPadding,
     RunningMoments,
     disable_dropout_in_model,
     generate_model_card,
+    get_comet_experiment_url,
     pad_to_length,
     peft_module_casting_to_bf16,
 )
