+++ trl/trainer/cpo_trainer.py
@@ -48,7 +48,6 @@ from transformers.utils import is_peft_available, is_torch_fx_proxy
 from transformers.utils.deprecation import deprecate_kwarg
 
 from ..data_utils import maybe_apply_chat_template, maybe_extract_prompt
-from ..utils import get_comet_experiment_url
 from .cpo_config import CPOConfig
 from .utils import (
     DPODataCollatorWithPadding,
@@ -56,6 +55,7 @@ from .utils import (
     add_eos_token_if_needed,
     disable_dropout_in_model,
     generate_model_card,
+    get_comet_experiment_url,
     pad_to_length,
     peft_module_casting_to_bf16,
 )
