+++ trl/trainer/orpo_trainer.py
@@ -52,7 +52,6 @@ from transformers.utils.deprecation import deprecate_kwarg
 
 from ..data_utils import maybe_apply_chat_template, maybe_extract_prompt
 from ..models import PreTrainedModelWrapper
-from ..utils import get_comet_experiment_url
 from .orpo_config import ORPOConfig
 from .utils import (
     DPODataCollatorWithPadding,
@@ -60,6 +59,7 @@ from .utils import (
     add_eos_token_if_needed,
     disable_dropout_in_model,
     generate_model_card,
+    get_comet_experiment_url,
     pad_to_length,
     peft_module_casting_to_bf16,
 )
