+++ trl/utils.py
@@ -15,9 +15,6 @@
 from dataclasses import dataclass
 from typing import Optional
 
-import pandas as pd
-from transformers import is_comet_available
-
 
 @dataclass
 class ScriptArguments:
@@ -46,37 +43,3 @@ class ScriptArguments:
     dataset_test_split: str = "test"
     gradient_checkpointing_use_reentrant: bool = False
     ignore_bias_buffers: bool = False
-
-
-def get_comet_experiment_url() -> Optional[str]:
-    """If Comet integration is enabled returns the URL of the current Comet experiment
-    or None if disabled or no Comet experiment is currently running."""
-    if not is_comet_available():
-        return None
-
-    import comet_ml
-
-    if comet_ml.get_running_experiment() is not None:
-        return comet_ml.get_running_experiment().url
-
-    return None
-
-
-def log_table_to_comet_experiment(name: str, table: pd.DataFrame) -> None:
-    """
-    If Comet integration is enabled logs a table to the Comet experiment if it is currently running.
-
-    Args:
-        name (`str`):
-            Table name.
-        table (`pd.DataFrame`):
-            The Pandas DataFrame containing the table to log.
-    """
-    if not is_comet_available():
-        return
-
-    import comet_ml
-
-    experiment = comet_ml.get_running_experiment()
-    if experiment is not None:
-        experiment.log_table(tabular_data=table, filename=name)
